+++ src/chunk_store/immutable.rs
@@ -7,7 +7,7 @@
 // permissions and limitations relating to use of the SAFE Network Software.
 
 use super::chunk::{Chunk, ChunkId};
-use sn_data_types::{Blob, BlobAddress};
+use sn_data_types::{Blob, BlobAddress, DataAddress};
 
 impl Chunk for Blob {
     type Id = BlobAddress;
@@ -19,4 +19,8 @@ impl Chunk for Blob {
     }
 }
 
-impl ChunkId for BlobAddress {}
+impl ChunkId for BlobAddress {
+    fn to_data_address(&self) -> DataAddress {
+        DataAddress::Blob(*self)
+    }
+}