+++ src/chunk_store/register.rs
@@ -7,7 +7,10 @@
 // permissions and limitations relating to use of the SAFE Network Software.
 
 use super::chunk::{Chunk, ChunkId};
-use sn_data_types::register::{Address, Register};
+use sn_data_types::{
+    register::{Address, Register},
+    DataAddress,
+};
 
 impl Chunk for Register {
     type Id = Address;
@@ -16,4 +19,8 @@ impl Chunk for Register {
     }
 }
 
-impl ChunkId for Address {}
+impl ChunkId for Address {
+    fn to_data_address(&self) -> DataAddress {
+        DataAddress::Register(*self)
+    }
+}